Add id validation in delete method in HivesController

+++ KatlaSport.WebApi/Controllers/HivesController.cs
@@ -70,8 +70,13 @@ namespace KatlaSport.WebApi.Controllers
         [SwaggerResponse(HttpStatusCode.Conflict)]
         [SwaggerResponse(HttpStatusCode.NotFound)]
         [SwaggerResponse(HttpStatusCode.InternalServerError)]
-        public async Task<IHttpActionResult> DeleteProductAsync([FromUri] int id)
+        public async Task<IHttpActionResult> DeleteHiveAsync([FromUri] int id)
         {
+            if (id <= 1)
+            {
+                return BadRequest();
+            }
+
             await _hiveService.DeleteHiveAsync(id);
             return ResponseMessage(Request.CreateResponse(HttpStatusCode.NoContent));
         }
